\name{MixtureDistribution-package}
\alias{MixtureDistribution-package}
\alias{MixtureDistribution}
\docType{package}
\title{
  Mixture Distribution
}
\description{
  Contains functions to operate with generic linear mixture distributions of R distribution functions.
}
\keyword{distribution}
